test: cover a skip on a book whose entries share a page
frontmatter_toc.epub opens on a cover with no TOC entry, then six front-matter
entries anchored inside one rendered page, then three chapters. That is the
shape from the report. The new case checks href and progression after every
skip; a TOC index alone cannot show whether the reader moved.

Adding it took epub_test.dart past 300 LOC, so the navigation groups moved to
epub_navigation_test.dart, and the wait helper both files use moved to
helpers/expect_eventually.dart.
