+++ flureadium/example/assets/pubs/README.md
@@ -19,6 +19,8 @@ Some ones prefixed with a number are self-produced by Nota.
 | sample_visual.divina      | divina            | Minimal visual narrative    |
 | tap_targets.epub          | epub              | Full-viewport link, then plain page |
 | fixed_layout.epub         | epub              | Single pre-paginated page   |
+| hierarchical_toc.epub     | epub              | Nested nav document         |
+| frontmatter_toc.epub      | epub              | Six front-matter entries on one page |
 
 Both are self-produced minimal fixtures for the tap chain, but they are consumed
 differently.
@@ -35,3 +37,12 @@ test: a synthesized tap reaches the native view and that script path never repor
 so it is exercised by hand through the app's **Open Fixed Layout** button as part of the
 `user | tap` row in `validators.conf`. See "What a synthesized tap can reach" in
 `docs/05-testing/integration-tests.md`.
+
+`frontmatter_toc.epub` drives the `navigation (frontmatter_toc)` group in
+`integration_test/epub_test.dart`. Its nav document anchors six front-matter
+entries — a title, a byline, an imprint line — inside one short document that
+renders as a single page, with the chapters at the same nesting depth in later
+documents. A cover resource sits ahead of it and carries no TOC entry, so the
+first skip resolves from "no current entry" exactly as the reported book did.
+Without the visibility walk in `TocSkipNavigationMixin`, the second skip targets
+an entry on the page already shown and the reader does not move.
