+++ flureadium/example/integration_test/helpers/expect_eventually.dart
@@ -1,0 +1,17 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'pump_until.dart';
+
+/// Pumps until [condition] holds, failing with [reason] when it never does.
+///
+/// [pumpUntil] reports a timeout in its return value, so every wait has to
+/// assert that value or a never-satisfied condition passes silently.
+Future<void> expectEventually(
+  WidgetTester tester,
+  bool Function() condition, {
+  required String reason,
+  Duration timeout = const Duration(seconds: 15),
+}) async {
+  final satisfied = await pumpUntil(tester, condition, timeout: timeout);
+  expect(satisfied, isTrue, reason: reason);
+}
